+++ src/ima-plugin.js
@@ -5,9 +5,8 @@ import videojs from 'video.js';
 import 'videojs-contrib-ads';
 import './ima-player.js';
 
-const registerPlugin = videojs.registerPlugin || videojs.plugin;
 // basic plugin is enough for this purpose
-registerPlugin('ima', function(options) {
+videojs.registerPlugin('ima', function(options) {
 	// inits contrib-ads asap
 	this.ads(Object.assign({
 		debug: options.debug||false,